data_cleansing/requirements.txt
pandas
psycopg2-binary
vaderSentiment
textblob